Adding multiple resources as part of controller
Created the folder structure and added some resource endpoints in the controller to view, add, update and delete contact information in the database.
Added Swagger documentation.
Integrated code with the embedded in-memory H2 database for testing.
Added JUnit test cases for some of the methods of the controller.
Added Input Validations and Exception Handling mechanisms to the resources